banking_api v0.1.0 BankingApi.Accounts

Accounts Context.

Este módulo é responsável por expor as funcionalidades relacionadas à acounts e users.

Resumidamente, o objetivo do context é desacoplar a responsabilidade de models e controllers.

Functions

create_user(params)

Cria um user, uma account(conta bancária) e faz a associação entre ambos no banco de dados

O argumento da função é:

params: Um map com as chaves de user struct a serem inseridos.

Há dois possíveis retornos para a função:

Uma tupla do tipo: {:ok, user, account}, onde user e account são structs

Uma tupla do tipo: {:error, changeset}

## Examples

iex> user = %{
  email: "email@email.com",
  first_name: "first_name",
  last_name: "last_name",
  password: "pwd",
  password_confirmation: "pwd"
}
iex> Accounts.create_user(user)
%{:ok, %User{...}, %Account{...}}

get!(id)

Busca por uma account.

O argumento desta função é:

id: O id da account em formato Integer

O retorno desta função é uma struct de account.

Examples

iex> Accounts.get!(10)
%Account{...}

get_user!(id)

Busca o user no banco de dados e vincula o resultado com a respectiva conta bancária.

Argumentos da função:

id: O id do usuário em formato Integet

O retorno da função é uma user struct carregada com a respectiva account struct.

Examples

iex> Accounts.get_user!(10)
%User{accounts: %Account{...}, ...}