banking_api v0.1.0 BankingApi.Accounts.Auth.Guardian

Módulo responsável por lidar com as funcionalidades de autenticação do Guardian.

Conceitos importantes sobre o guardian:

claims: informações sobre o token, como, tempo de duração, tipo de acesso, subject e o próprio token

resource: O que será autenticado. No nosso caso, o user.

token: É um hash muito grande que armazena as informações de todas as claims.

subject ou sub: É um identificador curto que é usado para identificar a resource

Functions

after_encode_and_sign(r, claims, token, _)

Callback implementation for Guardian.after_encode_and_sign/4.

after_sign_in(conn, r, t, c, o)

Callback implementation for Guardian.after_sign_in/5.

authenticate(email, password)

Faz a autenticação do usuário no sistema e cria o token de autenticação

Argumentos para a função:

email: email do usuário

senha: senha do usuário

Existem dois possíveis retornos da função:

{:error, :unauthorized}: Caso falhe a autenticação

{:ok, user, token}: Onde, user é uma user struct já com as informações de account e token é o token gerado pelo Guardian

before_sign_out(conn, location, opts)

Callback implementation for Guardian.before_sign_out/3.

build_claims(c, _, _)

Callback implementation for Guardian.build_claims/3.

config()

Specs

config() :: Keyword.t()

Fetches the configuration for this module

config(key, default \\ nil)

Specs

config(atom() | String.t(), any()) :: any()

Returns a resolved value of the configuration found at a key.

See Guardian.Config.resolve_value/1.

decode_and_verify(token, claims_to_check \\ %{}, opts \\ [])

Specs

decode_and_verify(
  Guardian.Token.token(),
  Guardian.Token.claims(),
  Guardian.options()
) :: {:ok, Guardian.Token.claims()} | {:error, any()}

Decodes and verifies a token using the configuration on the implementation module.

See Guardian.decode_and_verify/4

default_token_type()

Specs

default_token_type() :: String.t()

The default type of token for this module

encode_and_sign(resource, claims \\ %{}, opts \\ [])

Specs

encode_and_sign(any(), Guardian.Token.claims(), Guardian.options()) ::
  {:ok, Guardian.Token.token(), Guardian.Token.claims()} | {:error, any()}

Encodes the claims. See Guardian.encode_and_sign/4 for more information

exchange(token, from_type, to_type, opts \\ [])

Specs

exchange(
  token :: Guardian.Token.token(),
  from_type :: String.t() | [String.t(), ...],
  to_type :: String.t(),
  options :: Guardian.options()
) ::
  {:ok, {Guardian.Token.token(), Guardian.Token.claims()},
   {Guardian.Token.token(), Guardian.Token.claims()}}
  | {:error, any()}

Exchanges a token of one type for another.

See Guardian.exchange for more information

on_exchange(old_stuff, new_stuff, options)

Callback implementation for Guardian.on_exchange/3.

on_refresh(old_stuff, new_stuff, options)

Callback implementation for Guardian.on_refresh/3.

on_revoke(claims, token, options)

Callback implementation for Guardian.on_revoke/3.

on_verify(claims, token, options)

Callback implementation for Guardian.on_verify/3.

peek(token)

Specs

peek(String.t()) :: map()

Provides the content of the token but without verification of either the claims or the signature.

Claims will be present at the :claims key.

See Guardian.peek/2 for more information

refresh(old_token, opts \\ [])

Specs

refresh(Guardian.Token.token(), Guardian.options()) ::
  {:ok, {Guardian.Token.token(), Guardian.Token.claims()},
   {Guardian.Token.token(), Guardian.Token.claims()}}
  | {:error, any()}

Refresh a token.

See Guardian.refresh for more information

resource_from_claims(claims)

Busca o usuário no banco de dados a partir do subject.

O retorno da função é uma tupla, que contém o atom :ok e a user struct

resource_from_token(token, claims_to_check \\ %{}, opts \\ [])

Specs

resource_from_token(
  token :: Guardian.Token.token(),
  claims_to_check :: Guardian.Token.claims() | nil,
  opts :: Guardian.options()
) :: {:ok, Guardian.Token.resource(), Guardian.Token.claims()} | {:error, any()}

Fetch the resource and claims directly from a token

See Guardian.resource_from_token for more information

revoke(token, opts \\ [])

Specs

revoke(Guardian.Token.token(), Guardian.options()) ::
  {:ok, Guardian.Token.claims()} | {:error, any()}

Revoke a token.

See Guardian.revoke for more information

subject_for_token(user, claims)

Busca o subject do token do user.

verify_claims(claims, options)

Callback implementation for Guardian.verify_claims/2.