banking_api v0.1.0 BankingApi.Operations

Operations Context.

Este módulo é responsável por expor as funcionalidades que atuam sobre as operações transfer e withdraw.

Resumidamente, o objetivo de um context é desacoplar a responsabilidade de models e controllers.

Functions

is_transfering_to_same_account?(from, to_id)

Verifica se está transferindo para a mesma conta.

Os argumentos da função são:

from: account struct da conta que realizará a transferência

to_id: id da conta que receberá a transferência

Examples

iex> from = %Account{balance: "1000.00", user_id: "1"}
iex> Operations.is_transfering_to_same_account?(from, "1")
{:error, "Você não pode transferir para a sua própria conta"}

perform_operation(account, amount, atom)

Faz a soma de amount na conta bancária.

Os argumentos da função são:

account: É uma account struct

amount: É a quantidade a ser somada

O atom :sum.

O retorno da função é uma tupla contendo o status da operação de update no banco de dados.

transfer(from, to_id, amount)

Inicia a operação de transferência bancária.

Os argumentos desta função são:

from: É a account struct da conta pagadora.

to_id: O número da conta recebedora (id da conta).

amount: Quantidade a ser transferida.

Há possíveis retornos para a função. Eles podem ser:

{:ok, message}: Caso a transferência tenha ocorrido com sucesso

{:error, message}: Caso o saldo da conta seja menor do que o valor a ser transferido.

Qualquer retorno da função handle_feedback/2

Um dos possíveis motivos da operação não ser realizada (e retornar error), é se o valor amount a ser transferido é maior do que from.balance.

Examples

iex> from = %Account{balance: "1000.00", user_id: "1"}
iex> Operations.transfer(from, "2", "50.00")
{:ok, "Transferência realizada da conta 1 para a conta 2 no valor de R$ 50.00. O saldo de sua conta bancária é R$ 950.00"}

update_account(account, params)

Valida os dados de account antes que eles sejam inseridos no banco de dados.

Os argumentos da função são:

account: É uma struc de account, e representa a conta que será atualizada.

params: É um map, e contém os valores a serem atualizados.

withdraw(from, amount)

Inicia a operação de saque.

Os argumentos desta função são:

from: É a account struct de onde será feito saque.

amount: Quantidade a ser transferida.

Há possíveis retornos para a função. Eles podem ser:

{:ok, message}: Caso o saque tenha ocorrido com sucesso

{:error, message}: Caso o saldo da conta seja menor do que o valor de saque.

Qualquer retorno da função handle_feedback/2

Um dos possíveis motivos da operação não ser realizada, é se o valor amount é maior do que from.balance.

Examples

iex> from = %Account{balance: "1000.00", user_id: "1"}
iex> Operations.withdraw(from, "50.00")
{:ok, "Saque realizado no valor de R$ 50.00 da conta 1. O saldo de sua conta bancária é R$ 950.00"}