banking_api v0.1.0 BankingApi.Transactions

Transactions Context.

Este módulo é responsável por buscar no banco de dados as transações armazenadas na tabela transactions.

Resumidamente, o objetivo de um context é desacoplar a responsabilidade de models e controllers.

Functions

all()

Retorna as transações armazenadas no banco de dados, e a soma da quantia de todas as transações.

O retorno da função é um map contendo o total transacionado e a lista de transações: %{total: 1000, transactions: [%Transactions{}, ...]}

Examples

iex> all()
%{total: 1000, transactions: [%Transactions{}, ...]}

change_transaction(transaction, attrs \\ %{})

Returns an %Ecto.Changeset{} for tracking transaction changes.

Examples

iex> change_transaction(transaction)
%Ecto.Changeset{data: %Transaction{}}

create_transaction(attrs \\ %{})

Creates a transaction.

Examples

iex> create_transaction(%{field: value})
{:ok, %Transaction{}}

iex> create_transaction(%{field: bad_value})
{:error, %Ecto.Changeset{}}

day(date)

Retorna as transações ocorridas na data date e a quantia total transacionada na data.

O argumento da função é:

date: A data em formato string. A data deve seguir o formato: "yyyy-mm-dd".

O retorno da função é um map e segue o padrão: %{total: 1000, transactions: [%Transactions{}, ...]}

delete_transaction(transaction)

Deletes a transaction.

Examples

iex> delete_transaction(transaction)
{:ok, %Transaction{}}

iex> delete_transaction(transaction)
{:error, %Ecto.Changeset{}}

get_transaction!(id)

Gets a single transaction.

Raises Ecto.NoResultsError if the Transaction does not exist.

Examples

iex> get_transaction!(123)
%Transaction{}

iex> get_transaction!(456)
** (Ecto.NoResultsError)

insert_transaction(params)

Insere a transaction no banco de dados.

O argumento da função é o map params que contém os valores a serem inseridos na tabela.

O retorno da função é uma tupla, e retornará o atom :ok, caso a inserção tenha ocorrido com sucesso, ou :error, caso haja falha durante a inserção.

Examples

iex> insert_transaction(%{field: value})
{:ok, %Transaction{}}
iex> create_user(%{field: bad_value})
{:error, %Ecto.Changeset{}}

month(year, month)

Retorna as transações ocorridas no mês month do ano year e a quantia total transacionada na data.

Os argumentos da função são:

year: O ano em formato string. O ano deve seguir o formato: "yyyy".

month: O mês em formato string. O mês deve seguir o formato: "mm"

O retorno da função é um map e segue o padrão: %{total: 1000, transactions: [%Transactions{}, ...]}

Examples

iex> month(year, month) %{total: 1000, transactions: [%Transactions{}, ...]}

update_transaction(transaction, attrs)

Updates a transaction.

Examples

iex> update_transaction(transaction, %{field: new_value})
{:ok, %Transaction{}}

iex> update_transaction(transaction, %{field: bad_value})
{:error, %Ecto.Changeset{}}

validate_date(date, fail_message)

Chama a função BankingApi.Transactions.Helper.validate_date/2 que faz a validação de data.

Os argumentos da função são:

date: A data, em formato string, a ser validada

fail_message: A mensagem a ser exibida ao usuário caso a data passada esteja inválida

Os dois possíveis retornos da função são uma túpla contendo o atom :ok, caso a data seja válida, ou uma tupla contendo o atom :error e a mensagem fail_message, caso a data seja inválida

year(year)

Retorna as transações ocorridas no ano year e a quantia total transacionada na data.

O argumento da função é: year: O ano em formato string. O ano deve seguir o formato: "yyyy".

O retorno da função é um map contendo o total transacionado e a lista de transações: %{total: 1000, transactions: [%Transactions{}, ...]}

Examples

iex> year(year)
%{total: 1000, transactions: [%Transactions{}, ...]}