banking_api v0.1.0 BankingApi.Transactions.Helper

Modulo helper ao Transactions Context

Functions

create_report(transactions)

A função monta o que será o report de transações.

O argumento da função é:

transactions: Uma lista de transaction structs

O retorno da função é um map com duas keys:

total: O valor total transacionado

transactions: Uma lista de transaction structs

Examples

iex> transactions = [ %{
    value: "100",
    account_from: "1",
    type: "withdraw",
    account_to: "~",
    date: ~D[2019-10-30]
  }]
iex> Helper.create_report(transactions)
%{
  total: #Decimal<100>,
  transactions: [%{account_from: "1", account_to: "~", date: ~D[2019-10-30], type: "withdraw", value: "100"}]
}

do_query(start_date, end_date)

A função monta e executa a query que sera executada na tabela tranactions.

A consulta no banco de dados é feita baseada em um período de tempo. Portanto, os argumentos da função são:

start_date: O primeiro dia a ser levado em consideração para a consulta. A variável é do tipo string e deve seguir o formato "yyyy-mm-dd"

end_date: O último dia a ser levado em consideração para a consulta. A variável é do tipo string e deve seguir o formato "yyyy-mm-dd"

O retorno da função é um map com duas keys:

total: O valor total transacionado

transactions: Uma lista de transaction structs

Examples

iex> Helper.do_query("2020-01-01", "2020-12-30")
%{total: #Decimal<100>, transactions: [%Transactions{}, ...]}

list_transactions()

A função busca todos os registros da tabela transactions.

O retorno da função é uma lista de transaction structs

Examples

iex> Helper.list_transactions()
[%Transactions{}, ...]

query_by_day(date)

A função busca os registros da tabela transactions inseridos no dia especificado.

O argumento da função é:

date: A data em formato string. A data deve seguir o formato: "yyyy-mm-dd".

O retorno da função é uma lista de transaction structs

Examples

iex> Helper.query_by_day("2009-01-01")
[%Transactions{}, ...]

query_by_month(year, month)

A função busca os registros da tabela transactions inseridos no ano year e no mês month.

Os argumentos da função são:

year: O ano em formato integer. O ano deve seguir o formato: yyyy.

month: O mês em formato integer. O mês deve seguir o formato: mm

O retorno da função é um map que contém o total transacionado e uma lista de transaction structs

Examples

iex> Helper.query_by_month(2020, 10)
%{total: #Decimal<100>, transactions: [%Transactions{}, ...]}

query_by_year(year)

A função busca os registros da tabela transactions inseridos no ano year.

O argumento da função é:

year: O ano em formato integer. O ano deve seguir o formato: yyyy.

O retorno da função é um map que contém o total transacionado e uma lista de transaction structs

Examples

iex> Helper.query_by_year(2020)
%{total: #Decimal<100>, transactions: [%Transactions{}, ...]}

sum_amount(transactions)

A função é responsável por somar os valores de transaction.value de todas as structs passadas dentro da lista transactions

O argumento da função é a lista transactions que contém as structs de transaction

O retorno da função é o resultado da soma de todas as transaction.value da lista

validate_date(date, fail_message)

Valida se o ano, mês ou dia são uma string vazia ou contém o caractére "-".

Os argumentos da função são:

date: Pode ser o ano, o mês ou o dia, em formato string, a ser validado

fail_message: A mensagem a ser exibida ao usuário caso a data passada esteja inválida

Os dois possíveis retornos da função são uma túpla contendo o atom :ok, caso a data seja válida, ou uma tupla contendo o atom :error e a mensagem fail_message, caso a data seja inválida

Examples

iex> Helper.validate_date("2020", "this is an invalid year")
{:ok, "A data é válida"}

iex> Helper.validate_date("-2020", "this is an invalid year")
{:error, "this is an invalid year"}