banking_api v0.1.0 BankingApiWeb.FallbackController

O objetivo deste módulo é centralizar a manipulação de erros relacionadas ao BankingApiWeb.UserController.

Functions

call(conn, action)

A função call é responsável por retornar uma mensagem de erro ao usuário final quando há algum problema durante alguma operation, seja ela de transfer ou de withdraw. Os argumentos da função são as informações de connection conn e a mensagem de erro: message.