banking_api v0.1.0 BankingApiWeb.OperationController

Esse é o Controller responsável por todas as operações bancárias (transferência e saque).

Functions

transfer(conn, map)

Esta função é responsável por iniciar o processo de transferência bancária entre duas contas e renderizar o feedback sobre o status da operação para o usuário.

Os argumentos da função são:

conn: as informações da conexão

map: Um map que contém: 2.1. "to_account" => to: O número da conta recebedora (id da conta) 2.2. "amount" => amount: Valor a ser transferido

withdraw(conn, map)

Esta função é responsável por iniciar a operação de saque e renderizar o feedback sobre o status do processo para o usuário.

Os argumentos da função são:

conn: as informações da conexão

map: Um map que contém: 2.1. "amount" => amount: Valor a ser transferido