banking_api v0.1.0 BankingApiWeb.UserController

Esse é o User Controller. Este módulo é o primeiro contato entre as requisições e o back-end

Functions

show(conn, map)

Esta função é responsável por buscar um usuário pelo id e renderizar as informações deste usuário obtido. Os argumentos da função são as informações de conexão conn e o id do usuário que será resgatado

signup(conn, map)

A função signup é responsável por cadastrar um novo usuário no sistema e criar uma nova conta bancária para este usuário. O argumento user da função são os dados do novo usuário que será inserido no sistema. O retorno da função é a resposta ao usuário final que pode ser: uma mensagem informativa, caso o retorno de create_user tenha retornado {:ok, user, account}, ou uma mensagem de erro, caso o retorno de create_user tenha sido outro. Neste segundo caso, o Plug BankingApiWeb.FallbackController.call() é invocado, e lá e construida o feedback de erro para o usuário.