banking_api v0.1.0 BankingApi.Accounts

Este módulo representa a Accounts Context. Este módulo é responsável por expor as funcionalidades relacionadas à acounts e users. Resumidamente, o objetivo do context é desacoplar a responsabilidade de models e controllers.

Functions

create_user(params)

Função responsável por criar um user(usuário) e uma account(conta bancária) e a associação entre eles no banco de dados O argumento params são os dados do user a serem inseridos. O retorno da função é uma tupla de {:ok, user, account} ou {:error, changeset}

get!(id)

Esta função buscar uma account baseada em um id passado por parâmetro. O único parâmetro desta função é o account id. O retorno desta função é uma struct de account.

get_user!(id)

Esta função busca o user pelo id, vincula o resultado com a informação da conta bancária account. O argumento da função é o user id. O retorno da função é um map do usuário vinculado com a respectiva conta bancária.