banking_api v0.1.0 BankingApi.Operations

Este é o Operations Context. Este módulo é responsável por expor as funcionalidades relacionadas às operações de transfer e withdraw. Resumidamente, o objetivo de um context é desacoplar a responsabilidade de models e controllers.

Functions

perform_operation(account, amount, atom)

A função perform_operation/3 que recebe como último parâmetro o atom :sum é responsável por realizar a operação de soma na account passada por parâmetro. Os argumentos da função são uma struct account, a quantidade amount a ser somada e o atom :sum. O retorno da função é uma tupla contendo o status da operação de update no banco de dados.

perform_update(from, to_id, amount)

Esta função é responsável validar se a operação de subtração da conta pagadora (from) foi concluida, e se for, realizar a operação de soma na conta recebedora (to_id)

Os argumentos da função são:

from: É uma struct de account e representa a conta pagadora.

to_id: É o id da conta recebedora.

amount: Quantidade a ser transferida

transfer(from_id, to_id, amount)

Esta função é responsável por operar a transferência bancária entre duas contas.

Os argumentos desta função são:

from_id: O número da conta pagadora (id da conta).

to_id: O número da conta recebedora (id da conta).

amount: Quantidade a ser transferida.

O retorno da função é um map contendo o atom :error ou :ok. Representando o status da operação.

Um dos possíveis motivos da operação não ser realizada (e retornar um error), é se o valor amount a ser transferido é maior do que from.balance.

update_account(account, params)

Esta função é responsável por validar os dados de account antes que eles sejam inseridos no banco de dados.

Os argumentos da função são:

account: É uma struc de account, e representa a conta que será atualizada.

params: É um map, e contém os valores a serem atualizados.

withdraw(from_id, amount)

Esta função é responsável por operar a operação de saque.

Os argumentos desta função são:

from_id: Número da conta (id da conta) que realizará o saque.

amount: Quantidade a ser transferida.

O retorno da função é um map contendo o atom :error ou :ok. Representando o status da operação.

Um dos possíveis motivos da operação não ser realizada, é se o valor amount é maior do que from.balance.