banking_api v0.1.0 BankingApi.Transactions.Helper

Modulo helper ao Transactions Context

Functions

create_report(transactions)

A função monta o que será o report de transações.

O argumento da função é:

transactions: Uma lista de transaction structs

O retorno da função é um map com duas keys:

total: O valor total transacionado

transactions: Uma lista de transaction structs

do_query(start_date, end_date)

A função monta e executa a query que retornará as tranactions no banco de dados.

A consulta no banco de dados é feita baseada em um período de tempo. Portanto, os argumentos da função são:

start_date: O primeiro dia a ser levado em consideração para a consulta.

end_date: O último dia a ser levado em consideração para a consulta.

O retorno da função é um map com duas keys:

total: O valor total transacionado

transactions: Uma lista de transaction structs

list_transactions()

A função busca todos os registros da tabela transactions.

O retorno da função é uma lista de transaction structs

Examples

iex> list_transactions()
[%Transactions{}, ...]

query_by_day(date)

A função busca todos os registros da tabela transactions inseridos no ano year e no mês month.

O argumento da função é:

date: A data em formato string. A data deve seguir o formato: "yyyy-mm-dd".

O retorno da função é uma lista de transaction structs

Examples

iex> query_by_day(date)
[%Transactions{}, ...]

query_by_month(year, month)

A função busca todos os registros da tabela transactions inseridos no ano year e no mês month.

Os argumentos da função são:

year: O ano em formato integer. O ano deve seguir o formato: yyyy.

month: O mês em formato integer. O mês deve seguir o formato: mm

O retorno da função é uma lista de transaction structs

Examples

iex> query_by_month(year, month)
[%Transactions{}, ...]

query_by_year(year)

A função busca todos os registros da tabela transactions inseridos no ano year.

O argumento da função é: year: O ano em formato integer. O ano deve seguir o formato: yyyy.

O retorno da função é uma lista de transaction structs

Examples

iex> query_by_year(year)
[%Transactions{}, ...]

sum_amount(trasactions)

A função é responsável por somar os valores de transaction.value de todas as structs passadas dentro da lista transactions

O argumento da função é a lista transactions que contém as structs de transaction

O retorno da função é o resultado da soma de todas as transaction.value da lista

validate_date(date, fail_message)

Valida a data.

Os argumentos da função são:

date: A data, em formato string, a ser validada

fail_message: A mensagem a ser exibida ao usuário caso a data passada esteja inválida

Os dois possíveis retornos da função são uma túpla contendo o atom :ok, caso a data seja válida, ou uma tupla contendo o atom :error e a mensagem fail_message, caso a data seja inválida